feat: Add iframe support
Currently the code assumes the top-level `window` object owns the given node, but this isn't the case if the node resides within an iframe, targetted from the top-level window. This handles that particular case by using the `defaultView`/`ownerWindow` properties to get the parent window of the given node, and checks for `pageXOffset` instead of comparing directly to top-level `window` object.